Refactor: useGroups 생성 및 그에 따른 수정
